+++ src/frameworks/ruby-rails.ts
@@ -45,9 +45,17 @@ class RubyRailsFramework extends Framework {
     if (source === 'reference' && key.startsWith('.') && context.targetFile) {
       const root = path.resolve(Global.rootpath, Config.frameworksRubyRailsScopeRoot)
       let scope = path.relative(root, context.targetFile)
-      if (scope.endsWith('.html.erb'))
-        scope = scope.slice(0, -9)
-      scope = scope.replace(/(?:\/|\\)/g, '.')
+
+      scope = scope
+        // remove file extensions
+        .replace(/(?:\.html\.erb|\.html\.haml|\.haml)$/, '') 
+        // map path delimiter to dots
+        .replace(/(?:\/|\\)/g, '.') 
+        // omit the starting underscore on each file
+        .replace(/\._/g, '.') 
+        // remove the very starting underscore
+        .replace(/^_/g, '') 
+
       return scope + key
     }
     return key