+++ src/protocol/index.ts
@@ -73,6 +73,7 @@ export class Protocol implements Disposable {
       review: Config.reviewEnabled,
       locales,
       flags: locales.map(i => Config.tagSystem.getFlagName(i)),
+      showFlags: Config.showFlags,
       sourceLanguage: Config.sourceLanguage,
       displayLanguage: Config.displayLanguage,
       enabledFrameworks: Config.enabledFrameworks,