+++ src/tagSystems/bcp47.ts
@@ -1,18 +1,10 @@
 // @ts-ignore
 import bcp47 from 'bcp-47'
+import { Config } from '../core'
 import { BaseTagSystem } from './base'
 
 // https://tools.ietf.org/html/bcp47
 export class BCP47 extends BaseTagSystem {
-  // this maps language code to its default region
-  flagMapping: Record<string, string> = {
-    en: 'us',
-    zh: 'cn',
-    cs: 'cz',
-    da: 'dk',
-    el: 'gr',
-  }
-
   normalize(locale?: string, fallback = 'en', strict = false) {
     if (!locale)
       return fallback
@@ -26,8 +18,7 @@ export class BCP47 extends BaseTagSystem {
 
   toFlagname(locale: string) {
     const { region, language } = bcp47.parse(locale, { normalize: true, forgiving: true })
-    const flag = (region || language || '').toLowerCase()
-    return this.flagMapping[flag] || flag
+    return (region || Config.localeCountryMap[language] || '').toLowerCase()
   }
 
   lookup(locale: string) {